Just playing a game of squash the next undefined variable and seeing one piece at a time come back. :)
