+++ src/sketch/Custom.js
@@ -1,111 +1,4 @@
-var worldmapWord    = "worldmap";
-var customWord      = "custom";
-var orWord          = "or";
-var listOfWords     = ["Dawn", "Day", "Dusk", "Night", "Cloud", "Rain", "Snow", "Thund", "Hail", "Haze", "generate"];
-var timeOfDayScroll = 0;
 
-var nextButton = [
-    [0, 0, 0, 0, 0, 10, 10, 10, 20],
-    [0, 10, 20, 30, 40, 10, 20, 30, 20]
-];
-var checkMark = [
-    [0, 10, 20, 30, 40],
-    [20, 30, 20, 10, 0]
-]
-var infoMark = [
-    [10, 10, 20, 20, 20, 20, 20, 20, 30],
-    [30, 70, 10, 30, 40, 50, 60, 70, 70]
-]
-
-
-function Choose() {
-  stroke(30);
-
-  for (var i = 0; i < height; i = i + 10) {
-    line(0, i, width, i);
-  }
-
-  for (var i = 0; i < width; i = i + 10) {
-    line(i, 0, i, height);
-  }
-
-  fill(80);
-  noStroke();
-  stroke(0)
-
-  for (var i = 0; i < infoMark[0].length; i++) {
-    rect(
-      int(width / 10) * 10 - 80 + infoMark[0][i],
-      int(height / 10) * 10 - 120 + infoMark[1][i],
-      10,
-      10
-    );
-  }
-
-  for (var i = 1; i <= worldmapWord.length; i++) {
-    PixelLetter(
-      int(width / 10) * 10 / 2 + 70 - (i * 60),
-      int(height / 10) * 10 / 2 - 249,
-      worldmapWord.charAt(worldmapWord.length - i)
-    )
-  }
-
-  for (var i = 1; i <= customWord.length; i++) {
-    PixelLetter(
-      int(width / 10) * 10 / 2 + 350 - (i * 60),
-      int(height / 10) * 10 / 2 + 149,
-      customWord.charAt(customWord.length - i)
-    )
-  }
-
-  for (var i = 1; i <= orWord.length; i++) {
-    PixelLetter(
-      int(width / 10) * 10 / 2 + 70 - (i * 60),
-      int(height / 10) * 10 / 2 - 49,
-      orWord.charAt(orWord.length - i)
-    )
-  }
-
-  if (mouseX < width / 2 + 70 && mouseX > width / 2 - 410 && mouseY > height / 2 - 250 && mouseY < height / 2 - 180) {
-    fill(100);
-    stroke(80)
-    strokeWeight(1);
-
-    for (var i = 1; i <= worldmapWord.length; i++) {
-      PixelLetter(
-        int(width / 10) * 10 / 2 + 70 - (i * 60), int(height / 10) * 10 / 2 - 249, worldmapWord.charAt(worldmapWord.length - i))
-    }
-
-  } else if (mouseX < width / 2 + 350 && mouseX > width / 2 - 10 && mouseY > height / 2 + 150 && mouseY < height / 2 + 220) {
-
-    fill(100);
-    stroke(80)
-    strokeWeight(1);
-
-    for (var i = 1; i <= customWord.length; i++) {
-      PixelLetter(
-        int(width / 10) * 10 / 2 + 350 - (i * 60),
-        int(height / 10) * 10 / 2 + 149,
-        customWord.charAt(customWord.length - i)
-      )
-    }
-
-  } else if (mouseX < int(width / 10) * 10 - 50 && mouseX > int(width / 10) * 10 - 80 && mouseY > int(height / 10) * 10 - 120 && mouseY < int(height / 10) * 10 - 0) {
-
-    fill(100);
-    stroke(80)
-    strokeWeight(1);
-
-    for (var i = 0; i < infoMark[0].length; i++) {
-      rect(
-        int(width / 10) * 10 - 80 + infoMark[0][i],
-        int(height / 10) * 10 - 120 + infoMark[1][i],
-        10,
-        10
-      );
-    }
-  }
-}
 
 function Choice() {
     positionX = mouseX;
